requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.7
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
